Complete refactoring of plan route Test with Postman Changes to be committed: 	modified:   src/libs/db2/model/model.js 	modified:   src/router/controllers/PlanController.js 	new file:   src/router/controllers/PlanInvasivePlantController.js 	new file:   src/router/controllers/PlanManagementConsiderationController.js 	new file:   src/router/controllers/PlanMinisterIssueActionController.js 	modified:   src/router/controllers/PlanMinisterIssueController.js 	modified:   src/router/routes/plan.js

+++ src/libs/db2/model/model.js
@@ -199,15 +199,19 @@ export default class Model {
     }
   }
 
-  static removeById(db, id) {
+  static async removeById(db, id) {
     const where = {};
     where[this.primaryKey] = id;
 
     return this.remove(db, where);
   }
 
-  static remove(db, where) {
-    const results = db
+  static async remove(db, where) {
+    const exists = await this.find(db, where);
+    if (exists.length === 0) {
+      throw new Error(`Items not exist in [${this.table}] table with details: ${JSON.stringify(where)}`);
+    }
+    const results = await db
       .table(this.table)
       .where(where)
       .delete();
